Move sidebar construction from window.py into sidebar.build_sidebar
setup_sidebar's ~790-line body and the _expand_sidebar_toolbar_button
helper now live in sidebar.py next to the rows, colors, and DnD logic
they build on; MainWindow.setup_sidebar is a thin delegate. No behavior
change.
